+++ SparkEngine/Source/Engine/AI/BehaviorTree.h
@@ -62,602 +62,18 @@
  */
 
 #pragma once
-#include "BehaviorTreeTypes.h"
 
+// Subsystem headers
+#include "BehaviorTreeTypes.h"
+#include "BehaviorTreeNodes.h"
+
+#include <memory>
+#include <string>
+#include <vector>
 
 namespace Spark::AI
 {
 
-    // =============================================================================
-    // Composite Nodes
-    // =============================================================================
-
-    /**
- * @class SequenceNode
- * @brief Composite node that runs children left-to-right, stopping on first failure.
- *
- * A Sequence node is analogous to a logical AND:
- * - Ticks children in order.
- * - Returns **Failure** immediately if any child fails.
- * - Returns **Running** if a child is still running (resumes from that child next tick).
- * - Returns **Success** only when **all** children succeed.
- *
- * ### Typical use cases
- * - "Move to cover AND crouch AND wait 2 seconds" — all steps must complete.
- * - Precondition + action pairs: "CanSeeEnemy → Attack".
- *
- * @code
- *   auto seq = std::make_unique<SequenceNode>("PatrolStep");
- *   seq->AddChild(std::make_unique<ConditionNode>("HasWaypoint", hasWaypointFn));
- *   seq->AddChild(std::make_unique<ActionNode>("MoveToWaypoint", moveFn));
- *   seq->AddChild(std::make_unique<WaitNode>(1.0f));  // pause at waypoint
- * @endcode
- */
-    class SequenceNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Construct a SequenceNode with an optional display name.
-     * @param name  Human-readable name shown in the editor debugger. Default: "Sequence".
-     */
-        explicit SequenceNode(const std::string& name = "Sequence") : m_name(name) {}
-
-        /**
-     * @brief Append a child node to the end of the sequence.
-     *
-     * Children are ticked in the order they are added. Transfer ownership via
-     * `std::move()` when calling.
-     *
-     * @param child  Unique pointer to the child node to add.
-     */
-        void AddChild(std::unique_ptr<BTNode> child) { m_children.push_back(std::move(child)); }
-
-        /**
-     * @brief Tick children in order; fail fast on first failure.
-     *
-     * Maintains `m_currentChild` so that a Running child is resumed on the next tick
-     * rather than restarting from child 0.
-     *
-     * @param deltaTime  Frame delta time (seconds).
-     * @param blackboard AI working memory.
-     * @return           Running (waiting on a child), Failure (child failed), or
-     *                   Success (all children succeeded).
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            for (size_t i = m_currentChild; i < m_children.size(); ++i)
-            {
-                NodeStatus status = m_children[i]->Tick(deltaTime, blackboard);
-                if (status == NodeStatus::Running)
-                {
-                    m_currentChild = i;
-                    return m_status = NodeStatus::Running;
-                }
-                if (status == NodeStatus::Failure)
-                {
-                    m_currentChild = 0;
-                    return m_status = NodeStatus::Failure;
-                }
-            }
-            m_currentChild = 0;
-            return m_status = NodeStatus::Success;
-        }
-
-        /**
-     * @brief Reset the sequence and all its children to initial state.
-     *
-     * Resets `m_currentChild` to 0 so the sequence restarts from the beginning.
-     */
-        void Reset() override
-        {
-            BTNode::Reset();
-            m_currentChild = 0;
-            for (auto& child : m_children)
-                child->Reset();
-        }
-
-        const char* GetName() const override { return m_name.c_str(); }
-
-      private:
-        std::string m_name;
-        std::vector<std::unique_ptr<BTNode>> m_children;
-        /** @brief Index of the currently executing child; preserved across ticks for Running nodes. */
-        size_t m_currentChild = 0;
-    };
-
-    /**
- * @class SelectorNode
- * @brief Composite node that runs children left-to-right, stopping on first success.
- *
- * A Selector is analogous to a logical OR (priority selector):
- * - Ticks children in order (higher priority first).
- * - Returns **Success** immediately if any child succeeds.
- * - Returns **Running** if a child is still running (resumes from that child next tick).
- * - Returns **Failure** only when **all** children fail.
- *
- * ### Typical use cases
- * - Fallback chains: "TryAttack → TakeCover → Flee → Idle" (first available action wins).
- * - Priority overrides: highest-priority branch wins if its conditions are met.
- *
- * @code
- *   auto selector = std::make_unique<SelectorNode>("CombatSelector");
- *   selector->AddChild(std::make_unique<SequenceNode>("AttackIfInRange"));
- *   selector->AddChild(std::make_unique<SequenceNode>("ChaseTarget"));
- *   selector->AddChild(std::make_unique<ActionNode>("Idle", idleFn));
- * @endcode
- */
-    class SelectorNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Construct a SelectorNode with an optional display name.
-     * @param name  Human-readable name. Default: "Selector".
-     */
-        explicit SelectorNode(const std::string& name = "Selector") : m_name(name) {}
-
-        /**
-     * @brief Append a child at the lowest priority position.
-     * @param child  Unique pointer to the child node.
-     */
-        void AddChild(std::unique_ptr<BTNode> child) { m_children.push_back(std::move(child)); }
-
-        /**
-     * @brief Tick children in order; succeed fast on first success.
-     *
-     * @param deltaTime  Frame delta time (seconds).
-     * @param blackboard AI working memory.
-     * @return           Running, Success (first child succeeded), or Failure (all children failed).
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            for (size_t i = m_currentChild; i < m_children.size(); ++i)
-            {
-                NodeStatus status = m_children[i]->Tick(deltaTime, blackboard);
-                if (status == NodeStatus::Running)
-                {
-                    m_currentChild = i;
-                    return m_status = NodeStatus::Running;
-                }
-                if (status == NodeStatus::Success)
-                {
-                    m_currentChild = 0;
-                    return m_status = NodeStatus::Success;
-                }
-            }
-            m_currentChild = 0;
-            return m_status = NodeStatus::Failure;
-        }
-
-        /**
-     * @brief Reset the selector and all its children.
-     */
-        void Reset() override
-        {
-            BTNode::Reset();
-            m_currentChild = 0;
-            for (auto& child : m_children)
-                child->Reset();
-        }
-
-        const char* GetName() const override { return m_name.c_str(); }
-
-      private:
-        std::string m_name;
-        std::vector<std::unique_ptr<BTNode>> m_children;
-        /** @brief Index of the currently executing child. */
-        size_t m_currentChild = 0;
-    };
-
-    /**
- * @class ParallelNode
- * @brief Composite node that ticks **all** children simultaneously every frame.
- *
- * Unlike Sequence and Selector, a Parallel node ticks every child regardless of
- * their individual success or failure. The overall result is determined by a
- * configurable success policy.
- *
- * ### Success policies
- * - **RequireOne** – Succeeds as soon as any child succeeds; fails if all fail.
- * - **RequireAll** – Succeeds only when all children succeed; fails if any fail.
- *
- * ### Typical use cases
- * - Running multiple concurrent goals: "Patrol AND ScanForEnemies simultaneously".
- * - Interrupt-by-condition: "Parallel(MoveTo, Guard IsHealthLow → Flee)".
- *
- * @note Running a Parallel node is more expensive than Sequence/Selector because
- *       every child is ticked regardless of status. Prefer the latter when possible.
- */
-    class ParallelNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Success/failure evaluation policy for the Parallel node.
-     */
-        enum class Policy
-        {
-            RequireOne, ///< Succeed when at least one child succeeds.
-            RequireAll  ///< Succeed only when every child succeeds.
-        };
-
-        /**
-     * @brief Construct a ParallelNode with the given success policy and name.
-     * @param successPolicy  Determines when the node reports Success.
-     * @param name           Human-readable name. Default: "Parallel".
-     */
-        explicit ParallelNode(Policy successPolicy = Policy::RequireOne, const std::string& name = "Parallel")
-            : m_successPolicy(successPolicy), m_name(name)
-        {
-        }
-
-        /**
-     * @brief Append a child to the parallel group.
-     * @param child  Unique pointer to the child node.
-     */
-        void AddChild(std::unique_ptr<BTNode> child) { m_children.push_back(std::move(child)); }
-
-        /**
-     * @brief Tick all children; evaluate overall status via the configured policy.
-     *
-     * All children are ticked unconditionally. Success/failure counts are tallied
-     * and the overall result is derived from `m_successPolicy`.
-     *
-     * @param deltaTime  Frame delta time (seconds).
-     * @param blackboard AI working memory.
-     * @return           Running, Success, or Failure as determined by the policy.
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            int successCount = 0, failureCount = 0;
-
-            for (auto& child : m_children)
-            {
-                NodeStatus status = child->Tick(deltaTime, blackboard);
-                if (status == NodeStatus::Success)
-                    successCount++;
-                else if (status == NodeStatus::Failure)
-                    failureCount++;
-            }
-
-            if (m_successPolicy == Policy::RequireAll)
-            {
-                if (successCount == static_cast<int>(m_children.size()))
-                    return m_status = NodeStatus::Success;
-                if (failureCount > 0)
-                    return m_status = NodeStatus::Failure;
-            }
-            else
-            {
-                if (successCount > 0)
-                    return m_status = NodeStatus::Success;
-                if (failureCount == static_cast<int>(m_children.size()))
-                    return m_status = NodeStatus::Failure;
-            }
-            return m_status = NodeStatus::Running;
-        }
-
-        /**
-     * @brief Reset all children in the parallel group.
-     */
-        void Reset() override
-        {
-            BTNode::Reset();
-            for (auto& child : m_children)
-                child->Reset();
-        }
-
-        const char* GetName() const override { return m_name.c_str(); }
-
-      private:
-        Policy m_successPolicy;
-        std::string m_name;
-        std::vector<std::unique_ptr<BTNode>> m_children;
-    };
-
-    // =============================================================================
-    // Decorator Nodes
-    // =============================================================================
-
-    /**
- * @class InverterNode
- * @brief Decorator that negates its child's Success/Failure result.
- *
- * Passes `Running` through unchanged. Converts `Success` to `Failure` and
- * `Failure` to `Success`. Useful for building "NOT" conditions:
- * e.g. "Succeed if player is NOT in sight".
- *
- * @code
- *   // Succeed when the player is NOT visible
- *   auto notVisible = std::make_unique<InverterNode>(
- *       std::make_unique<ConditionNode>("IsPlayerVisible", isVisibleFn));
- * @endcode
- */
-    class InverterNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Construct the inverter wrapping the given child node.
-     * @param child  The single child whose result will be negated.
-     */
-        explicit InverterNode(std::unique_ptr<BTNode> child) : m_child(std::move(child)) {}
-
-        /**
-     * @brief Tick the child and invert its Success/Failure result.
-     * @param deltaTime  Frame delta time (seconds).
-     * @param blackboard AI working memory.
-     * @return           Inverted result (Running unchanged, Success↔Failure).
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            NodeStatus status = m_child->Tick(deltaTime, blackboard);
-            if (status == NodeStatus::Success)
-                return m_status = NodeStatus::Failure;
-            if (status == NodeStatus::Failure)
-                return m_status = NodeStatus::Success;
-            return m_status = NodeStatus::Running;
-        }
-
-        const char* GetName() const override { return "Inverter"; }
-
-      private:
-        /** @brief The single child node whose result is inverted. */
-        std::unique_ptr<BTNode> m_child;
-    };
-
-    /**
- * @class RepeaterNode
- * @brief Decorator that re-ticks its child a fixed number of times (or infinitely).
- *
- * After each non-Running result from the child, the child is reset and re-ticked
- * on the next frame. When `repeatCount > 0`, the Repeater returns Success after
- * that many completions. When `repeatCount <= 0` (default: -1) the child repeats
- * indefinitely, which is useful for looping patrol behaviors.
- *
- * @note An infinite repeater never returns Success or Failure; it always returns
- *       Running. Wrap it in a Selector or use an InterruptCondition to escape.
- *
- * @code
- *   // Repeat a patrol loop forever
- *   auto loopPatrol = std::make_unique<RepeaterNode>(
- *       std::move(patrolSequence), -1);
- * @endcode
- */
-    class RepeaterNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Construct a RepeaterNode.
-     * @param child        The child to repeat.
-     * @param repeatCount  Number of repetitions. Negative value = infinite. Default: -1.
-     */
-        RepeaterNode(std::unique_ptr<BTNode> child, int repeatCount = -1)
-            : m_child(std::move(child)), m_repeatCount(repeatCount)
-        {
-        }
-
-        /**
-     * @brief Tick the child; reset and loop when it completes.
-     *
-     * @param deltaTime  Frame delta time (seconds).
-     * @param blackboard AI working memory.
-     * @return           Running while repeating, Success after `repeatCount` completions.
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            NodeStatus status = m_child->Tick(deltaTime, blackboard);
-            if (status == NodeStatus::Running)
-                return m_status = NodeStatus::Running;
-
-            m_currentCount++;
-            if (m_repeatCount > 0 && m_currentCount >= m_repeatCount)
-                return m_status = NodeStatus::Success;
-
-            m_child->Reset();
-            return m_status = NodeStatus::Running;
-        }
-
-        /**
-     * @brief Reset the repeater and its child, clearing the repeat counter.
-     */
-        void Reset() override
-        {
-            BTNode::Reset();
-            m_currentCount = 0;
-            m_child->Reset();
-        }
-
-        const char* GetName() const override { return "Repeater"; }
-
-      private:
-        std::unique_ptr<BTNode> m_child;
-        /** @brief Target number of repetitions (-1 = infinite). */
-        int m_repeatCount;
-        /** @brief Number of completions so far in the current execution. */
-        int m_currentCount = 0;
-    };
-
-    // =============================================================================
-    // Leaf Nodes
-    // =============================================================================
-
-    /**
- * @class ActionNode
- * @brief Leaf node that executes a custom user-supplied function each tick.
- *
- * ActionNode is the primary way to integrate game logic into a behavior tree.
- * The action function receives `deltaTime` and a reference to the Blackboard, and
- * returns a NodeStatus to signal whether the action is still ongoing, succeeded, or failed.
- *
- * ### Writing action functions
- * @code
- *   auto moveToTarget = std::make_unique<ActionNode>("MoveToTarget",
- *       [this](float dt, Blackboard& bb) -> NodeStatus
- *       {
- *           XMFLOAT3 target = bb.Get<XMFLOAT3>("targetPos");
- *           float dist = MathUtils::Distance(m_agent->GetPosition(), target);
- *           if (dist < 0.5f) return NodeStatus::Success;
- *           m_agent->MoveTowards(target, dt);
- *           return NodeStatus::Running;
- *       });
- * @endcode
- *
- * @note Prefer short, single-responsibility action functions. Complex multi-step
- *       processes should be modeled as a SequenceNode of simpler ActionNodes.
- */
-    class ActionNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Function signature for the action callback.
-     *
-     * @param deltaTime  Frame time in seconds.
-     * @param blackboard Reference to the tree's shared blackboard.
-     * @return           NodeStatus indicating ongoing, success, or failure.
-     */
-        using ActionFunc = std::function<NodeStatus(float, Blackboard&)>;
-
-        /**
-     * @brief Construct an ActionNode with a name and callback.
-     * @param name    Human-readable identifier shown in the editor debugger.
-     * @param action  The function executed each tick.
-     */
-        ActionNode(const std::string& name, ActionFunc action) : m_name(name), m_action(std::move(action)) {}
-
-        /**
-     * @brief Invoke the action function and return its result.
-     * @param deltaTime  Frame delta time (seconds).
-     * @param blackboard AI working memory.
-     * @return           The status returned by the action function.
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            return m_status = m_action(deltaTime, blackboard);
-        }
-
-        const char* GetName() const override { return m_name.c_str(); }
-
-      private:
-        std::string m_name;
-        ActionFunc m_action;
-    };
-
-    /**
- * @class ConditionNode
- * @brief Leaf node that tests a boolean predicate against the Blackboard.
- *
- * Returns `NodeStatus::Success` if the condition function returns `true`, and
- * `NodeStatus::Failure` otherwise. Never returns Running since the check is
- * instantaneous.
- *
- * Conditions are typically used as the first child of a SequenceNode (a guard),
- * or as a child of an InverterNode to build negated guards.
- *
- * @code
- *   // "Is the enemy closer than 10 metres?"
- *   auto isInRange = std::make_unique<ConditionNode>("IsInAttackRange",
- *       [this](const Blackboard& bb) {
- *           float dist = bb.Get<float>("distToTarget", FLT_MAX);
- *           return dist < 10.0f;
- *       });
- * @endcode
- */
-    class ConditionNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Function signature for the condition predicate.
-     *
-     * @param blackboard Const reference to the tree's blackboard.
-     * @return           `true` → Success; `false` → Failure.
-     */
-        using ConditionFunc = std::function<bool(const Blackboard&)>;
-
-        /**
-     * @brief Construct a ConditionNode with a name and predicate.
-     * @param name       Human-readable identifier.
-     * @param condition  Predicate function evaluated each tick.
-     */
-        ConditionNode(const std::string& name, ConditionFunc condition)
-            : m_name(name), m_condition(std::move(condition))
-        {
-        }
-
-        /**
-     * @brief Evaluate the condition and return Success or Failure.
-     *
-     * @param deltaTime  Unused for conditions (included for interface consistency).
-     * @param blackboard AI working memory read by the condition function.
-     * @return           Success if condition returns true, Failure otherwise.
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            return m_status = m_condition(blackboard) ? NodeStatus::Success : NodeStatus::Failure;
-        }
-
-        const char* GetName() const override { return m_name.c_str(); }
-
-      private:
-        std::string m_name;
-        ConditionFunc m_condition;
-    };
-
-    /**
- * @class WaitNode
- * @brief Leaf node that pauses execution for a specified duration.
- *
- * Returns `NodeStatus::Running` until the accumulated elapsed time reaches
- * `duration`, then returns `NodeStatus::Success` and resets the timer for
- * subsequent calls.
- *
- * Commonly used to add delays to patrol sequences (pause at waypoints) or
- * to implement attack cooldowns within a behavior branch.
- *
- * @code
- *   // Pause at a waypoint for 3 seconds before continuing the patrol
- *   auto pause = std::make_unique<WaitNode>(3.0f);
- * @endcode
- */
-    class WaitNode : public BTNode
-    {
-      public:
-        /**
-     * @brief Construct a WaitNode.
-     * @param duration  How many seconds to wait before returning Success.
-     */
-        explicit WaitNode(float duration) : m_duration(duration) {}
-
-        /**
-     * @brief Accumulate elapsed time; return Success when the duration is reached.
-     * @param deltaTime  Time since last tick (seconds).
-     * @param blackboard AI working memory (unused by WaitNode but required by interface).
-     * @return           Running while waiting, Success when duration elapsed.
-     */
-        NodeStatus Tick(float deltaTime, Blackboard& blackboard) override
-        {
-            m_elapsed += deltaTime;
-            if (m_elapsed >= m_duration)
-            {
-                m_elapsed = 0.0f;
-                return m_status = NodeStatus::Success;
-            }
-            return m_status = NodeStatus::Running;
-        }
-
-        /**
-     * @brief Reset the elapsed timer to zero.
-     */
-        void Reset() override
-        {
-            BTNode::Reset();
-            m_elapsed = 0.0f;
-        }
-        const char* GetName() const override { return "Wait"; }
-
-      private:
-        /** @brief Target wait duration in seconds. */
-        float m_duration;
-        /** @brief Accumulated elapsed time since last reset (seconds). */
-        float m_elapsed = 0.0f;
-    };
 
     // =============================================================================
     // Behavior Tree
