+++ SparkEngine/Source/Graphics/RenderGraph.h
@@ -1,86 +1,33 @@
 /**
  * @file RenderGraph.h
- * @brief Declarative render graph / frame graph system for SparkEngine
+ * @brief Declarative render graph / frame graph system for SparkEngine (umbrella header)
  * @author Spark Engine Team
  * @date 2026
  *
- * Provides a declarative render graph that allows the rendering pipeline to be
- * expressed as a directed acyclic graph (DAG) of render passes with explicit
- * resource dependencies. The graph is compiled once per frame structure change,
- * producing an optimized execution order with automatic resource lifetime
- * management, barrier insertion, and transient resource aliasing.
- *
- * Key components:
- * - RenderGraphResource: Handle to a transient or imported GPU resource
- * - RenderGraphPass: A single render/compute/copy operation in the graph
- * - RenderGraphBuilder: DSL used inside pass setup lambdas to declare I/O
- * - RenderGraph: Top-level object that owns passes and resources
- * - RenderGraphBlackboard: Type-erased data sharing between passes
- *
- * Usage:
- * @code
- *   RenderGraph graph("MainFrame");
- *
- *   struct GBufferData
- *   {
- *       RenderGraphResource albedo;
- *       RenderGraphResource normals;
- *       RenderGraphResource depth;
- *   };
- *
- *   auto& gbufferData = graph.GetBlackboard().Add<GBufferData>();
- *
- *   graph.AddPass("GBufferPass", RenderGraphPassType::Graphics,
- *       [&](RenderGraphBuilder& builder)
- *       {
- *           RenderGraphTextureDesc desc;
- *           desc.width = 1920;
- *           desc.height = 1080;
- *           desc.format = RenderTargetFormat::RGBA8_UNORM;
- *           gbufferData.albedo = builder.Create("GBuffer_Albedo", desc);
- *           gbufferData.normals = builder.Create("GBuffer_Normals", desc);
- *
- *           RenderGraphTextureDesc depthDesc;
- *           depthDesc.width = 1920;
- *           depthDesc.height = 1080;
- *           depthDesc.format = RenderTargetFormat::D32_FLOAT;
- *           gbufferData.depth = builder.Create("GBuffer_Depth", depthDesc);
- *
- *           gbufferData.albedo = builder.Write(gbufferData.albedo);
- *           gbufferData.normals = builder.Write(gbufferData.normals);
- *           gbufferData.depth = builder.Write(gbufferData.depth);
- *       },
- *       [=](const RenderGraphResourceRegistry& registry)
- *       {
- *           // Actual rendering code using registry to resolve handles
- *       });
- *
- *   graph.Compile();
- *   graph.Execute();
- * @endcode
+ * This is the umbrella header for the render graph subsystem. It includes:
+ * - RenderGraphTypes.h   — Resource handles, descriptors, enums, registry, stats
+ * - RenderGraphPass.h    — Pass and builder classes
+ * - RenderGraphBlackboard.h — Type-erased inter-pass data sharing
+ * - RenderGraph class    — Top-level graph object (defined below)
  *
  * @see GraphicsEngine, RenderTarget, RenderTargetFormat
  */
 
 #pragma once
 
-#include "../Core/Platform.h"
-#include "RenderTarget.h"
+// Subsystem headers
+#include "RenderGraphTypes.h"
+#include "RenderGraphPass.h"
+#include "RenderGraphBlackboard.h"
 
 #include <algorithm>
-#include <any>
 #include <cassert>
 #include <chrono>
-#include <cstdint>
 #include <functional>
 #include <memory>
-#include <mutex>
-#include <optional>
 #include <queue>
 #include <sstream>
 #include <string>
-#include <typeindex>
-#include <typeinfo>
 #include <unordered_map>
 #include <unordered_set>
 #include <vector>
@@ -94,601 +41,6 @@ class GraphicsEngine;
 
 namespace Spark::Graphics
 {
-
-    // ============================================================================
-    // Forward Declarations
-    // ============================================================================
-
-    class RenderGraph;
-    class RenderGraphBuilder;
-    class RenderGraphPass;
-    class RenderGraphResourceRegistry;
-
-    // ============================================================================
-    // RenderGraphResource — Typed handle to a graph-managed resource
-    // ============================================================================
-
-    /**
- * @brief Opaque handle representing a resource within the render graph.
- *
- * Resources are never accessed directly through this handle. Instead, the
- * handle is resolved to concrete GPU objects (render targets, buffers) via
- * RenderGraphResourceRegistry during pass execution.
- *
- * A resource handle carries a version number so that write-after-write
- * hazards produce distinct handles, enabling correct dependency tracking.
- */
-    struct RenderGraphResource
-    {
-        static constexpr uint32_t INVALID_INDEX = ~0u;
-
-        uint32_t index = INVALID_INDEX; ///< Index into the graph's resource array
-        uint32_t version = 0;           ///< Incremented on each write
-
-        constexpr bool IsValid() const { return index != INVALID_INDEX; }
-
-        constexpr bool operator==(const RenderGraphResource& other) const
-        {
-            return index == other.index && version == other.version;
-        }
-
-        constexpr bool operator!=(const RenderGraphResource& other) const { return !(*this == other); }
-    };
-
-    /**
- * @brief Hash support for RenderGraphResource so it can be used in
- *        unordered containers.
- */
-    struct RenderGraphResourceHash
-    {
-        size_t operator()(const RenderGraphResource& r) const
-        {
-            return std::hash<uint64_t>{}((static_cast<uint64_t>(r.index) << 32) | r.version);
-        }
-    };
-
-    // ============================================================================
-    // Resource Descriptors
-    // ============================================================================
-
-    /**
- * @brief Descriptor for a texture resource managed by the render graph.
- *
- * Mirrors the fields in RenderTargetDesc but is decoupled so that the graph
- * can reason about resources without creating GPU objects up front.
- */
-    struct RenderGraphTextureDesc
-    {
-        uint32_t width = 1;
-        uint32_t height = 1;
-        uint32_t depth = 1;       ///< For 3D textures; 1 for 2D
-        uint32_t arraySize = 1;   ///< Texture array layers
-        uint32_t mipLevels = 1;   ///< Mip chain length
-        uint32_t sampleCount = 1; ///< MSAA samples
-        RenderTargetFormat format = RenderTargetFormat::RGBA8_UNORM;
-        RenderTargetUsage usage = RenderTargetUsage::RenderTarget | RenderTargetUsage::ShaderResource;
-        XMFLOAT4 clearColor = {0.0f, 0.0f, 0.0f, 1.0f};
-        float clearDepth = 1.0f;
-        uint8_t clearStencil = 0;
-
-        /**
-     * @brief Compute the approximate memory footprint in bytes.
-     *
-     * This is used by the graph compiler for aliasing decisions.
-     */
-        size_t EstimateMemoryBytes() const
-        {
-            // Rough per-pixel byte count based on format
-            size_t bpp = 4; // default RGBA8
-            switch (format)
-            {
-            case RenderTargetFormat::RGBA32_FLOAT:
-                bpp = 16;
-                break;
-            case RenderTargetFormat::RGBA16_FLOAT:
-                bpp = 8;
-                break;
-            case RenderTargetFormat::RG32_FLOAT:
-                bpp = 8;
-                break;
-            case RenderTargetFormat::RG16_FLOAT:
-                bpp = 4;
-                break;
-            case RenderTargetFormat::R32_FLOAT:
-            case RenderTargetFormat::D32_FLOAT:
-                bpp = 4;
-                break;
-            case RenderTargetFormat::R16_FLOAT:
-            case RenderTargetFormat::D16_UNORM:
-                bpp = 2;
-                break;
-            case RenderTargetFormat::R8_UNORM:
-                bpp = 1;
-                break;
-            case RenderTargetFormat::R11G11B10_FLOAT:
-                bpp = 4;
-                break;
-            case RenderTargetFormat::RGB10A2_UNORM:
-                bpp = 4;
-                break;
-            case RenderTargetFormat::D24_UNORM_S8_UINT:
-                bpp = 4;
-                break;
-            default:
-                bpp = 4;
-                break;
-            }
-            return static_cast<size_t>(width) * height * depth * arraySize * sampleCount * bpp;
-        }
-    };
-
-    /**
- * @brief Descriptor for a buffer resource managed by the render graph.
- */
-    struct RenderGraphBufferDesc
-    {
-        size_t sizeBytes = 0; ///< Total buffer size
-        size_t stride = 0;    ///< Element stride (for structured buffers)
-        RenderTargetUsage usage = RenderTargetUsage::ShaderResource;
-
-        size_t EstimateMemoryBytes() const { return sizeBytes; }
-    };
-
-    // ============================================================================
-    // Resource Lifetime — distinguishes transient from imported resources
-    // ============================================================================
-
-    /**
- * @brief Indicates whether a resource is graph-managed (transient) or
- *        externally owned (imported).
- */
-    enum class RenderGraphResourceLifetime
-    {
-        Transient, ///< Created and destroyed within a single graph execution
-        Imported   ///< Externally owned; the graph does not manage its lifetime
-    };
-
-    /**
- * @brief The kind of GPU resource a handle points to.
- */
-    enum class RenderGraphResourceType
-    {
-        Texture,
-        Buffer
-    };
-
-    // ============================================================================
-    // Internal Resource Node
-    // ============================================================================
-
-    /**
- * @brief Internal bookkeeping for a single resource inside the graph.
- *
- * Users never interact with this directly; they use RenderGraphResource
- * handles and resolve them through the registry at execution time.
- */
-    struct RenderGraphResourceNode
-    {
-        std::string name;
-        RenderGraphResourceType type = RenderGraphResourceType::Texture;
-        RenderGraphResourceLifetime lifetime = RenderGraphResourceLifetime::Transient;
-
-        /// Descriptor — exactly one of these is populated depending on type.
-        RenderGraphTextureDesc textureDesc;
-        RenderGraphBufferDesc bufferDesc;
-
-        /// Reference count: incremented by each pass that reads or writes.
-        uint32_t refCount = 0;
-
-        /// The pass that first creates / writes this resource (producer).
-        uint32_t producerPass = RenderGraphResource::INVALID_INDEX;
-
-        /// Execution-time concrete object. Populated during Execute().
-        std::shared_ptr<RenderTarget> physicalTexture;
-
-        /// Aliasing: if non-null, this resource shares memory with another.
-        uint32_t aliasTarget = RenderGraphResource::INVALID_INDEX;
-
-        /// First and last pass indices that reference this resource (inclusive).
-        /// Used for lifetime analysis and aliasing.
-        uint32_t firstUsePass = RenderGraphResource::INVALID_INDEX;
-        uint32_t lastUsePass = 0;
-
-        /**
-     * @brief Imported texture pointer. Non-null only when lifetime == Imported.
-     */
-        RenderTarget* importedTexture = nullptr;
-    };
-
-    // ============================================================================
-    // RenderGraphPassType
-    // ============================================================================
-
-    /**
- * @brief Classifies a render pass for scheduling purposes.
- */
-    enum class RenderGraphPassType
-    {
-        Graphics,    ///< Rasterization pass (vertex/pixel shaders, render targets)
-        Compute,     ///< Compute-shader dispatch
-        Copy,        ///< Resource copy / blit
-        AsyncCompute ///< Compute work eligible for async compute queue
-    };
-
-    // ============================================================================
-    // RenderGraphResourceRegistry — resolves handles during execution
-    // ============================================================================
-
-    /**
- * @brief Passed to the execute callback of each pass so that it can
- *        resolve RenderGraphResource handles to concrete GPU objects.
- */
-    class RenderGraphResourceRegistry
-    {
-      public:
-        explicit RenderGraphResourceRegistry(const std::vector<RenderGraphResourceNode>& resources)
-            : m_resources(resources)
-        {
-        }
-
-        /**
-     * @brief Resolve a resource handle to its physical RenderTarget.
-     * @param handle The resource handle obtained during pass setup.
-     * @return Non-owning pointer to the render target, or nullptr if
-     *         the resource is not backed by a texture.
-     */
-        RenderTarget* GetTexture(RenderGraphResource handle) const
-        {
-            if (!handle.IsValid() || handle.index >= m_resources.size())
-            {
-                return nullptr;
-            }
-            const auto& node = m_resources[handle.index];
-            if (node.lifetime == RenderGraphResourceLifetime::Imported)
-            {
-                return node.importedTexture;
-            }
-            return node.physicalTexture.get();
-        }
-
-        /**
-     * @brief Get the texture descriptor for a resource.
-     */
-        const RenderGraphTextureDesc& GetTextureDesc(RenderGraphResource handle) const
-        {
-            return m_resources[handle.index].textureDesc;
-        }
-
-        /**
-     * @brief Get the buffer descriptor for a resource.
-     */
-        const RenderGraphBufferDesc& GetBufferDesc(RenderGraphResource handle) const
-        {
-            return m_resources[handle.index].bufferDesc;
-        }
-
-      private:
-        const std::vector<RenderGraphResourceNode>& m_resources;
-    };
-
-    // ============================================================================
-    // RenderGraphPass — a single node in the pass DAG
-    // ============================================================================
-
-    /**
- * @brief Represents a single render, compute, or copy pass in the graph.
- *
- * Passes are created via RenderGraph::AddPass(). Each pass records which
- * resources it reads and writes so that the compiler can determine
- * execution order and insert barriers.
- */
-    class RenderGraphPass
-    {
-      public:
-        using SetupCallback = std::function<void(RenderGraphBuilder&)>;
-        using ExecuteCallback = std::function<void(const RenderGraphResourceRegistry&)>;
-
-        /**
-     * @brief Construct a render graph pass.
-     * @param name Human-readable name (shown in debug visualization).
-     * @param type The kind of GPU work this pass performs.
-     * @param index Unique index within the owning RenderGraph.
-     */
-        RenderGraphPass(const std::string& name, RenderGraphPassType type, uint32_t index)
-            : m_name(name), m_type(type), m_index(index)
-        {
-        }
-
-        // -- Accessors -----------------------------------------------------------
-
-        const std::string& GetName() const { return m_name; }
-        RenderGraphPassType GetType() const { return m_type; }
-        uint32_t GetIndex() const { return m_index; }
-        bool HasSideEffects() const { return m_hasSideEffects; }
-        bool IsCulled() const { return m_culled; }
-
-        const std::vector<RenderGraphResource>& GetReads() const { return m_reads; }
-        const std::vector<RenderGraphResource>& GetWrites() const { return m_writes; }
-        const std::vector<RenderGraphResource>& GetCreates() const { return m_creates; }
-
-        // -- Mutators (used by RenderGraphBuilder and compiler) -------------------
-
-        void AddRead(RenderGraphResource r) { m_reads.push_back(r); }
-        void AddWrite(RenderGraphResource r) { m_writes.push_back(r); }
-        void AddCreate(RenderGraphResource r) { m_creates.push_back(r); }
-        void MarkSideEffects() { m_hasSideEffects = true; }
-        void SetCulled(bool culled) { m_culled = culled; }
-
-        void SetExecuteCallback(ExecuteCallback callback) { m_executeCallback = std::move(callback); }
-
-        /**
-     * @brief Run the pass's execute callback.
-     *
-     * Called by RenderGraph::Execute() after all resources have been
-     * allocated and barriers inserted.
-     */
-        void Execute(const RenderGraphResourceRegistry& registry) const
-        {
-            if (m_executeCallback)
-            {
-                m_executeCallback(registry);
-            }
-        }
-
-      private:
-        std::string m_name;
-        RenderGraphPassType m_type;
-        uint32_t m_index;
-
-        std::vector<RenderGraphResource> m_reads;
-        std::vector<RenderGraphResource> m_writes;
-        std::vector<RenderGraphResource> m_creates;
-
-        bool m_hasSideEffects = false;
-        bool m_culled = false;
-
-        ExecuteCallback m_executeCallback;
-    };
-
-    // ============================================================================
-    // RenderGraphBuilder — pass-setup DSL
-    // ============================================================================
-
-    /**
- * @brief Builder interface provided to pass setup lambdas.
- *
- * The builder is the sole mechanism through which a pass declares its
- * resource dependencies. After the setup lambda returns, the graph owns
- * all dependency information and uses it for compilation.
- */
-    class RenderGraphBuilder
-    {
-      public:
-        RenderGraphBuilder(RenderGraph& graph, RenderGraphPass& pass) : m_graph(graph), m_pass(pass) {}
-
-        // -- Resource creation ---------------------------------------------------
-
-        /**
-     * @brief Create a new transient texture resource.
-     * @param name Human-readable name for debugging.
-     * @param desc Texture descriptor.
-     * @return Handle to the newly created resource.
-     */
-        RenderGraphResource Create(const std::string& name, const RenderGraphTextureDesc& desc);
-
-        /**
-     * @brief Create a new transient buffer resource.
-     * @param name Human-readable name for debugging.
-     * @param desc Buffer descriptor.
-     * @return Handle to the newly created resource.
-     */
-        RenderGraphResource Create(const std::string& name, const RenderGraphBufferDesc& desc);
-
-        // -- Resource access declarations ----------------------------------------
-
-        /**
-     * @brief Declare that this pass reads from a resource.
-     *
-     * The returned handle has the same index and version as the input,
-     * but the dependency is recorded so the compiler orders this pass
-     * after the producer.
-     *
-     * @param resource Handle obtained from a prior Create() or Write().
-     * @return The same handle (pass-through for chaining convenience).
-     */
-        RenderGraphResource Read(RenderGraphResource resource);
-
-        /**
-     * @brief Declare that this pass writes to a resource.
-     *
-     * Returns a new handle with an incremented version. Subsequent
-     * passes that want to read this resource must use the returned handle.
-     *
-     * @param resource Handle to write.
-     * @return New versioned handle representing the resource after this write.
-     */
-        RenderGraphResource Write(RenderGraphResource resource);
-
-        // -- Side effects --------------------------------------------------------
-
-        /**
-     * @brief Mark this pass as having side effects (e.g., presenting to
-     *        the swap chain, writing to disk).
-     *
-     * Side-effect passes are never culled during dead-code elimination.
-     */
-        RenderGraphBuilder& SideEffect()
-        {
-            m_pass.MarkSideEffects();
-            return *this;
-        }
-
-        // -- Resource aliasing ---------------------------------------------------
-
-        /**
-     * @brief Hint that two resources with non-overlapping lifetimes may
-     *        share the same physical memory.
-     *
-     * The compiler will attempt to honor this hint when allocating
-     * transient resources. Both resources must have compatible descriptors.
-     *
-     * @param from The resource whose memory may be reused.
-     * @param to   The resource that should alias from's memory.
-     */
-        void Alias(RenderGraphResource from, RenderGraphResource to);
-
-      private:
-        RenderGraph& m_graph;
-        RenderGraphPass& m_pass;
-    };
-
-    // ============================================================================
-    // RenderGraphBlackboard — type-erased inter-pass data sharing
-    // ============================================================================
-
-    /**
- * @brief Type-erased storage for passing structured data between passes.
- *
- * Each data type is stored at most once, keyed by its std::type_index.
- * Passes write data during setup and read it during execution.
- *
- * @code
- *   struct LightingData { RenderGraphResource shadowMap; };
- *   auto& data = blackboard.Add<LightingData>();
- *   data.shadowMap = builder.Create("ShadowMap", desc);
- *   // Later, in another pass:
- *   const auto& data = blackboard.Get<LightingData>();
- * @endcode
- */
-    class RenderGraphBlackboard
-    {
-      public:
-        RenderGraphBlackboard() = default;
-        ~RenderGraphBlackboard() = default;
-
-        // Non-copyable, movable
-        RenderGraphBlackboard(const RenderGraphBlackboard&) = delete;
-        RenderGraphBlackboard& operator=(const RenderGraphBlackboard&) = delete;
-        RenderGraphBlackboard(RenderGraphBlackboard&&) noexcept = default;
-        RenderGraphBlackboard& operator=(RenderGraphBlackboard&&) noexcept = default;
-
-        /**
-     * @brief Add or replace data of type T.
-     *
-     * Constructs a T in-place. If a T already exists, it is replaced.
-     *
-     * @tparam T   Data type (must be movable).
-     * @tparam Args Constructor argument types.
-     * @param args  Forwarded to T's constructor.
-     * @return Reference to the stored T.
-     */
-        template <typename T, typename... Args> T& Add(Args&&... args)
-        {
-            auto key = std::type_index(typeid(T));
-            m_storage[key] = std::make_any<T>(std::forward<Args>(args)...);
-            return std::any_cast<T&>(m_storage[key]);
-        }
-
-        /**
-     * @brief Retrieve mutable reference to stored data of type T.
-     * @throws std::bad_any_cast if T has not been added.
-     */
-        template <typename T> T& Get()
-        {
-            auto key = std::type_index(typeid(T));
-            auto it = m_storage.find(key);
-            if (it == m_storage.end())
-            {
-                throw std::bad_any_cast();
-            }
-            return std::any_cast<T&>(it->second);
-        }
-
-        /**
-     * @brief Retrieve const reference to stored data of type T.
-     * @throws std::bad_any_cast if T has not been added.
-     */
-        template <typename T> const T& Get() const
-        {
-            auto key = std::type_index(typeid(T));
-            auto it = m_storage.find(key);
-            if (it == m_storage.end())
-            {
-                throw std::bad_any_cast();
-            }
-            return std::any_cast<const T&>(it->second);
-        }
-
-        /**
-     * @brief Try to retrieve data of type T without throwing.
-     * @return Pointer to the stored T, or nullptr if not present.
-     */
-        template <typename T> T* TryGet()
-        {
-            auto key = std::type_index(typeid(T));
-            auto it = m_storage.find(key);
-            if (it == m_storage.end())
-            {
-                return nullptr;
-            }
-            return std::any_cast<T>(&it->second);
-        }
-
-        /**
-     * @brief Try to retrieve const data of type T without throwing.
-     */
-        template <typename T> const T* TryGet() const
-        {
-            auto key = std::type_index(typeid(T));
-            auto it = m_storage.find(key);
-            if (it == m_storage.end())
-            {
-                return nullptr;
-            }
-            return std::any_cast<const T>(&it->second);
-        }
-
-        /**
-     * @brief Check whether data of type T exists.
-     */
-        template <typename T> bool Has() const { return m_storage.count(std::type_index(typeid(T))) > 0; }
-
-        /**
-     * @brief Remove data of type T.
-     */
-        template <typename T> void Remove() { m_storage.erase(std::type_index(typeid(T))); }
-
-        /**
-     * @brief Remove all stored data.
-     */
-        void Clear() { m_storage.clear(); }
-
-      private:
-        std::unordered_map<std::type_index, std::any> m_storage;
-    };
-
-    // ============================================================================
-    // Render Graph Statistics
-    // ============================================================================
-
-    /**
- * @brief Statistics collected during graph compilation and execution.
- */
-    struct RenderGraphStats
-    {
-        uint32_t totalPasses = 0;        ///< Total declared passes
-        uint32_t culledPasses = 0;       ///< Passes removed by dead-code elimination
-        uint32_t executedPasses = 0;     ///< Passes actually executed
-        uint32_t totalResources = 0;     ///< Total declared resources
-        uint32_t transientResources = 0; ///< Graph-managed transient resources
-        uint32_t importedResources = 0;  ///< Externally imported resources
-        uint32_t aliasedResources = 0;   ///< Resources sharing physical memory
-        size_t peakTransientMemory = 0;  ///< Peak transient memory in bytes
-        size_t savedByAliasing = 0;      ///< Memory saved via aliasing in bytes
-        float compileTimeMs = 0.0f;      ///< Time spent in Compile()
-        float executeTimeMs = 0.0f;      ///< Time spent in Execute()
-        uint32_t asyncComputePasses = 0; ///< Passes scheduled on async compute
-    };
 
     // ============================================================================
     // RenderGraph — top-level graph object
