+++ SparkEngine/Source/Physics/PhysicsSystem.h
@@ -1,56 +1,13 @@
 /**
  * @file PhysicsSystem.h
- * @brief Complete physics integration system for Spark Engine.
+ * @brief PhysicsSystem class and umbrella include for the physics subsystem
  * @author Spark Engine Team
  * @date 2025
  *
- * @details
- * PhysicsSystem wraps the **Bullet Physics 3** dynamics library, exposing a
- * DirectX Math-native API that integrates cleanly with the rest of Spark Engine.
- * All DirectX `XMFLOAT3` / `XMMATRIX` types are transparently converted to and
- * from Bullet's `btVector3` / `btTransform` internally.
- *
- * ## Architecture overview
- *
- * | Class              | Responsibility                                                    |
- * |--------------------|-------------------------------------------------------------------|
- * | PhysicsBody        | Wrapper around a `btRigidBody` with a DirectX Math transform API |
- * | PhysicsConstraint  | Wrapper around a `btTypedConstraint` (hinge, slider, fixed, …)   |
- * | PhysicsSystem      | World manager: lifecycle, simulation step, queries, callbacks     |
- *
- * ## Supported features
- *
- * - **Body types**: Static (immovable), Kinematic (animated by game logic),
- *   Dynamic (fully simulated).
- * - **Collision shapes**: Box, Sphere, Capsule, Cylinder, Cone, Mesh,
- *   ConvexHull, Heightfield, Compound.
- * - **Constraints**: Hinge, Slider, Fixed (more via `btTypedConstraint` directly).
- * - **Queries**: Raycast (single/all), SphereOverlap, BoxOverlap.
- * - **Callbacks**: Collision and trigger enter/exit events via `std::function`.
- * - **Materials**: Named PhysicsMaterial presets (friction, restitution, damping).
- * - **Debug rendering**: Optional Bullet debug-draw overlay via a custom DebugDrawer.
- * - **Console integration**: 12+ runtime commands for inspecting/modifying physics.
- *
- * ## Typical usage
- * @code
- *   PhysicsSystem physics;
- *   if (FAILED(physics.Initialize())) return;
- *
- *   // Create a dynamic box
- *   PhysicsBodyDesc desc;
- *   desc.type             = PhysicsBodyType::Dynamic;
- *   desc.position         = {0, 5, 0};
- *   desc.mass             = 10.0f;
- *   desc.shape.type       = CollisionShapeType::Box;
- *   desc.shape.dimensions = {1, 1, 1};
- *   auto box = physics.CreateBody(desc);
- *
- *   // Game loop
- *   while (running) {
- *       physics.Update(deltaTime);
- *       renderTransform = box->GetTransform();
- *   }
- * @endcode
+ * This is the umbrella header for the physics subsystem. It includes:
+ * - PhysicsTypes.h — Enums, descriptors, collision types, materials
+ * - PhysicsBody.h  — PhysicsBody and PhysicsConstraint wrapper classes
+ * - PhysicsSystem  — Central manager for the Bullet Physics dynamics world (below)
  *
  * @note PhysicsSystem is **not thread-safe**. Call all public methods from the
  *       main game thread. The physics simulation itself runs on the calling thread
@@ -58,9 +15,12 @@
  */
 
 #pragma once
+
+// Subsystem headers
+#include "PhysicsBody.h"
+
 #include "../Core/Platform.h"
 
-#include "Utils/Assert.h"
 #ifdef SPARK_PLATFORM_WINDOWS
 #include <DirectXMath.h>
 #endif // SPARK_PLATFORM_WINDOWS
@@ -78,545 +38,10 @@ class btDefaultCollisionConfiguration;
 class btCollisionDispatcher;
 class btDbvtBroadphase;
 class btSequentialImpulseConstraintSolver;
-class btRigidBody;
 class btCollisionShape;
 class btMotionState;
-class btTypedConstraint;
 class btGhostPairCallback;
 class btTriangleMesh;
-
-// Physics enums, data structs, and utility functions
-#include "PhysicsTypes.h"
-
-/**
- * @class PhysicsBody
- * @brief DirectX Math-native wrapper around a Bullet `btRigidBody`.
- *
- * PhysicsBody is the primary handle callers use to interact with a single
- * rigid body in the simulation. It abstracts away Bullet's coordinate system
- * and type conversions, presenting all transforms and velocities as DirectX
- * `XMFLOAT3` / `XMMATRIX` values.
- *
- * ### Ownership
- * PhysicsBody instances are created and owned by PhysicsSystem via `shared_ptr`.
- * Callers receive a `shared_ptr<PhysicsBody>` from CreateBody(); keeping it alive
- * keeps the body active. Dropping the last reference triggers cleanup, but for
- * deterministic removal call `PhysicsSystem::RemoveBody()` explicitly.
- *
- * ### Simulation loop integration
- * After each PhysicsSystem::Update() call, Dynamic bodies have new positions and
- * rotations. Sync your rendered transform by calling GetTransform() once per frame:
- * @code
- *   physics.Update(dt);
- *   XMMATRIX world = body->GetTransform();
- *   renderable.SetWorldMatrix(world);
- * @endcode
- *
- * ### Kinematic bodies
- * For animated/scripted objects that should push other bodies around, set
- * `type = Kinematic`. Drive them by calling SetPosition() or SetTransform()
- * each frame before Physics::Update().
- */
-class PhysicsBody
-{
-  public:
-    /**
-     * @brief Construct a PhysicsBody wrapping the given Bullet rigid body.
-     *
-     * Called internally by PhysicsSystem::CreateBody(). Do not construct manually.
-     *
-     * @param desc        Descriptor used to create this body (stored for later queries).
-     * @param bulletBody  Bullet rigid body pointer (owned by the Bullet dynamics world;
-     *                    must outlive this PhysicsBody wrapper).
-     */
-    PhysicsBody(const PhysicsBodyDesc& desc, btRigidBody* bulletBody);
-
-    /**
-     * @brief Destructor. Does not remove the body from the Bullet world.
-     *
-     * Call PhysicsSystem::RemoveBody() before the last shared_ptr reference is
-     * dropped to ensure the body is properly unregistered from the dynamics world.
-     */
-    ~PhysicsBody();
-
-    // =========================================================================
-    // Transform
-    // =========================================================================
-
-    /**
-     * @brief Get the current world-space position of the body's center of mass.
-     *
-     * For Dynamic bodies this reflects the position after the most recent
-     * PhysicsSystem::Update() call. For Static bodies the position never changes.
-     *
-     * @return  World-space position in metres.
-     */
-    XMFLOAT3 GetPosition() const;
-
-    /**
-     * @brief Teleport the body to a new world-space position.
-     *
-     * For Dynamic bodies this should be used sparingly (e.g. level teleporters)
-     * as abrupt position changes can cause tunnelling and interpenetration.
-     * For Kinematic bodies, call this every frame to drive the animation.
-     *
-     * @param position  New world-space position (metres).
-     */
-    void SetPosition(const XMFLOAT3& position);
-
-    /**
-     * @brief Get the current world-space rotation as Euler angles (degrees).
-     *
-     * The rotation is extracted from the body's internal quaternion and converted
-     * to (pitch, yaw, roll) in degrees. Prefer GetTransform() if you need the full
-     * 4×4 matrix to avoid gimbal-lock issues.
-     *
-     * @return  Euler angles in degrees (X=pitch, Y=yaw, Z=roll).
-     */
-    XMFLOAT3 GetRotation() const;
-
-    /**
-     * @brief Set the world-space rotation from Euler angles (degrees).
-     *
-     * Converts (pitch, yaw, roll) to a quaternion internally. Useful for scripted
-     * or editor-driven orientation changes.
-     *
-     * @param rotation  Euler angles in degrees (X=pitch, Y=yaw, Z=roll).
-     */
-    void SetRotation(const XMFLOAT3& rotation);
-
-    /**
-     * @brief Get the body's full world-space transform as a 4×4 matrix.
-     *
-     * Returns a combined rotation + translation matrix suitable for use as a
-     * world matrix in a DirectX rendering pipeline. The matrix is column-major
-     * (consistent with DirectX Math conventions).
-     *
-     * @return  World-space transform matrix (rotation + translation; no scale).
-     */
-    XMMATRIX GetTransform() const;
-
-    /**
-     * @brief Set the body's full world-space transform from a 4×4 matrix.
-     *
-     * Decomposes the matrix into a translation and rotation, discarding any scale
-     * component (Bullet does not support per-body scaling). For Kinematic bodies,
-     * call this each frame to match the animated position/orientation.
-     *
-     * @param transform  World-space transform matrix. Scale is ignored.
-     */
-    void SetTransform(const XMMATRIX& transform);
-
-    // =========================================================================
-    // Velocity
-    // =========================================================================
-
-    /**
-     * @brief Get the body's current linear (translational) velocity.
-     *
-     * Only meaningful for Dynamic bodies; Static and Kinematic bodies always
-     * report zero or the velocity explicitly set via SetLinearVelocity().
-     *
-     * @return  Linear velocity in metres per second.
-     */
-    XMFLOAT3 GetLinearVelocity() const;
-
-    /**
-     * @brief Directly set the body's linear velocity, overriding any accumulated forces.
-     *
-     * Useful for snapping a body to a specific speed (e.g. setting a projectile's
-     * initial velocity). For continuous thrust, prefer ApplyForce() each frame.
-     *
-     * @param velocity  New linear velocity in metres per second.
-     */
-    void SetLinearVelocity(const XMFLOAT3& velocity);
-
-    /**
-     * @brief Get the body's current angular (rotational) velocity.
-     *
-     * Returned in radians per second around each world axis.
-     *
-     * @return  Angular velocity (radians/second) around X, Y, Z world axes.
-     */
-    XMFLOAT3 GetAngularVelocity() const;
-
-    /**
-     * @brief Directly set the body's angular velocity.
-     *
-     * Overwrites any spin accumulated from torques or collisions. Useful for
-     * controlling a thrown object's spin speed.
-     *
-     * @param velocity  New angular velocity in radians per second.
-     */
-    void SetAngularVelocity(const XMFLOAT3& velocity);
-
-    // =========================================================================
-    // Forces and impulses
-    // =========================================================================
-
-    /**
-     * @brief Apply a continuous force to the body (accumulates over the time step).
-     *
-     * Forces are integrated over the physics time step: `Δvelocity = force / mass × Δt`.
-     * Call this every frame for sustained effects like jet thrust or wind.
-     * The accumulated force is cleared automatically after each simulation step.
-     *
-     * @param force        Force vector in Newtons (world space).
-     * @param relativePos  Local-space offset from the center of mass where the force
-     *                     is applied. Non-zero values generate torque in addition to
-     *                     linear force. Defaults to {0,0,0} (center of mass).
-     */
-    void ApplyForce(const XMFLOAT3& force, const XMFLOAT3& relativePos = {0, 0, 0});
-
-    /**
-     * @brief Apply an instantaneous impulse to the body.
-     *
-     * An impulse produces an immediate velocity change: `Δvelocity = impulse / mass`.
-     * Use this for single-frame effects like explosions, jump boosts, or hit reactions.
-     *
-     * @param impulse      Impulse vector in Newton-seconds (N·s) (world space).
-     * @param relativePos  Local-space application offset. Non-zero values generate an
-     *                     angular impulse in addition to the linear one.
-     *                     Defaults to {0,0,0} (center of mass).
-     */
-    void ApplyImpulse(const XMFLOAT3& impulse, const XMFLOAT3& relativePos = {0, 0, 0});
-
-    /**
-     * @brief Apply a continuous torque to the body (accumulates over the time step).
-     *
-     * Causes angular acceleration: `Δω = torque / inertia × Δt`. Like ApplyForce(),
-     * the accumulated torque is reset after each simulation step.
-     *
-     * @param torque  Torque vector in Newton-metres (world space). The direction of
-     *                the vector is the axis of rotation; the magnitude is the strength.
-     */
-    void ApplyTorque(const XMFLOAT3& torque);
-
-    /**
-     * @brief Apply an instantaneous angular impulse to the body.
-     *
-     * Produces an immediate angular velocity change. Equivalent to a rotational
-     * version of ApplyImpulse(). Useful for imparting spin on impact.
-     *
-     * @param torque  Angular impulse in Newton-metre-seconds (N·m·s) (world space).
-     */
-    void ApplyTorqueImpulse(const XMFLOAT3& torque);
-
-    // =========================================================================
-    // Properties
-    // =========================================================================
-
-    /**
-     * @brief Get the body's mass in kilograms.
-     *
-     * Returns the effective mass used by the simulation. For Static bodies this
-     * is always 0 regardless of what was specified in the descriptor.
-     *
-     * @return  Mass in kilograms (0 for Static bodies).
-     */
-    float GetMass() const;
-
-    /**
-     * @brief Change the body's mass at runtime.
-     *
-     * Causes Bullet to recompute inertia tensors. Valid for Dynamic bodies; setting
-     * mass to 0 effectively converts the body to Static.
-     *
-     * @param mass  New mass in kilograms. Must be ≥ 0.
-     */
-    void SetMass(float mass);
-
-    /**
-     * @brief Get the body type (Static, Kinematic, or Dynamic).
-     * @return  PhysicsBodyType enum value as set in the creation descriptor.
-     */
-    PhysicsBodyType GetType() const { return m_desc.type; }
-
-    /**
-     * @brief Get the current physics material (const reference).
-     * @return  Const reference to the body's PhysicsMaterial.
-     */
-    const PhysicsMaterial& GetMaterial() const { return m_desc.material; }
-
-    /**
-     * @brief Replace the body's physics material at runtime.
-     *
-     * Updates friction, restitution, and damping on the underlying `btRigidBody`.
-     * Takes effect on the next simulation step.
-     *
-     * @param material  New material properties to apply.
-     */
-    void SetMaterial(const PhysicsMaterial& material);
-
-    // =========================================================================
-    // State
-    // =========================================================================
-
-    /**
-     * @brief Activate or deactivate the body in the simulation.
-     *
-     * Inactive bodies are excluded from collision detection and dynamics, making
-     * deactivation useful for pooled objects. Reactivate with `SetActive(true)` before
-     * reusing.
-     *
-     * @param active  `true` to activate; `false` to deactivate (sleep).
-     */
-    void SetActive(bool active);
-
-    /**
-     * @brief Check whether the body is currently active in the simulation.
-     * @return  `true` if the body is participating in the physics step; `false` if sleeping.
-     */
-    bool IsActive() const;
-
-    /**
-     * @brief Switch the body between Dynamic and Kinematic modes at runtime.
-     *
-     * Setting `kinematic = true` reconfigures the body as Kinematic: it will no
-     * longer respond to forces or gravity, but other Dynamic bodies will still
-     * respond to contacts with it.
-     *
-     * @param kinematic  `true` to make the body Kinematic; `false` to make it Dynamic.
-     */
-    void SetKinematic(bool kinematic);
-
-    /**
-     * @brief Check whether the body is currently Kinematic.
-     * @return  `true` if Kinematic; `false` if Dynamic or Static.
-     */
-    bool IsKinematic() const;
-
-    /**
-     * @brief Enable or disable trigger mode at runtime.
-     *
-     * A trigger body generates overlap callbacks (via PhysicsSystem::SetTriggerCallback())
-     * but applies no contact forces. Switching from solid to trigger (or vice versa)
-     * modifies the body's collision flags on the Bullet rigid body.
-     *
-     * @param trigger  `true` to enable trigger mode; `false` for solid collision.
-     */
-    void SetTrigger(bool trigger);
-
-    /**
-     * @brief Check whether trigger mode is enabled.
-     * @return  `true` if this body is a trigger volume; `false` if it is a solid collider.
-     */
-    bool IsTrigger() const;
-
-    // =========================================================================
-    // Collision filtering
-    // =========================================================================
-
-    /**
-     * @brief Set the collision group bitmask for this body.
-     *
-     * The body only collides with another body when:
-     *   `(this->group & other->mask) != 0 && (other->group & this->mask) != 0`.
-     *
-     * Default group: 1 (bit 0). Group 0 is reserved (no collisions at all).
-     *
-     * @param group  Bitmask identifying which "team" this body belongs to.
-     */
-    void SetCollisionGroup(uint16_t group);
-
-    /**
-     * @brief Set the collision mask for this body.
-     *
-     * The mask defines which groups this body will respond to. Bits set to 1
-     * mean "collide with bodies in that group".
-     *
-     * Default mask: 0xFFFF (collide with everything).
-     *
-     * @param mask  Bitmask of groups this body will collide with.
-     */
-    void SetCollisionMask(uint16_t mask);
-
-    /**
-     * @brief Get the current collision group bitmask.
-     * @return  16-bit collision group mask.
-     */
-    uint16_t GetCollisionGroup() const { return m_collisionGroup; }
-
-    /**
-     * @brief Get the current collision filter mask.
-     * @return  16-bit collision filter mask.
-     */
-    uint16_t GetCollisionMask() const { return m_collisionMask; }
-
-    // =========================================================================
-    // User data
-    // =========================================================================
-
-    /**
-     * @brief Attach a caller-owned data pointer to this body.
-     *
-     * The pointer is stored without any ownership semantics — the body never deletes
-     * it. Typically used to link the body back to a game object or ECS entity so that
-     * collision callbacks can identify and manipulate the owning object.
-     *
-     * @param data  Opaque pointer to caller-managed data. May be `nullptr`.
-     */
-    void SetUserData(void* data) { m_desc.userData = data; }
-
-    /**
-     * @brief Retrieve the previously set user data pointer.
-     * @return  Pointer set by SetUserData(), or `nullptr` if none was set.
-     */
-    void* GetUserData() const { return m_desc.userData; }
-
-    /**
-     * @brief Get the human-readable debug name of this body.
-     * @return  Name string as provided in PhysicsBodyDesc::name at creation time.
-     */
-    const std::string& GetName() const { return m_desc.name; }
-
-    // =========================================================================
-    // Internal / low-level access
-    // =========================================================================
-
-    /**
-     * @brief Get the underlying Bullet rigid body pointer.
-     *
-     * Use this for advanced Bullet operations not exposed by the PhysicsBody API.
-     * Handle with care — modifications to the Bullet body can desynchronize the
-     * PhysicsBody's cached state.
-     *
-     * @return  Non-owning pointer to the underlying `btRigidBody`.
-     */
-    btRigidBody* GetBulletBody() const { return m_bulletBody; }
-
-    /**
-     * @brief Get the creation descriptor for this body.
-     * @return  Const reference to the PhysicsBodyDesc used to create this body.
-     */
-    const PhysicsBodyDesc& GetDesc() const { return m_desc; }
-
-    // =========================================================================
-    // Console integration
-    // =========================================================================
-
-    /**
-     * @brief Return a formatted summary string for the debug console.
-     *
-     * Includes name, type, position, velocity, mass, material properties, and
-     * active/trigger flags. Bound to the console command `physics body info <name>`.
-     *
-     * @return  Human-readable multi-line info string.
-     */
-    std::string GetInfo() const;
-
-    /**
-     * @brief Set a named floating-point property via the debug console.
-     *
-     * Accepted property names: `"mass"`, `"friction"`, `"restitution"`,
-     * `"linearDamping"`, `"angularDamping"`.
-     *
-     * @param property  Property name (case-sensitive).
-     * @param value     New value to assign.
-     */
-    void Console_SetProperty(const std::string& property, float value);
-
-    /**
-     * @brief Apply a force impulse via the debug console.
-     *
-     * Equivalent to `ApplyImpulse({x, y, z})` — useful for testing physics
-     * interactions at runtime without modifying source code.
-     *
-     * @param x  Force X component (world space, N·s).
-     * @param y  Force Y component (world space, N·s).
-     * @param z  Force Z component (world space, N·s).
-     */
-    void Console_ApplyForce(float x, float y, float z);
-
-    // =========================================================================
-    // Interpolation
-    // =========================================================================
-
-    /**
-     * @brief Get the interpolated position between previous and current physics state.
-     *
-     * Used for smooth rendering at variable framerates when physics runs at a fixed
-     * timestep. When alpha is 0 the previous position is returned; when alpha is 1
-     * the current position is returned.
-     *
-     * @param alpha  Interpolation factor in [0, 1].
-     * @return       Interpolated world-space position.
-     */
-    XMFLOAT3 GetInterpolatedPosition(float alpha) const;
-
-    /**
-     * @brief Get the interpolated transform between previous and current physics state.
-     *
-     * Lerps position and Slerps rotation for smooth rendering between fixed physics steps.
-     *
-     * @param alpha  Interpolation factor in [0, 1].
-     * @return       Interpolated world-space transform matrix.
-     */
-    XMMATRIX GetInterpolatedTransform(float alpha) const;
-
-    /**
-     * @brief Snapshot the current state as the "previous" state for interpolation.
-     *
-     * Called internally by PhysicsSystem before each fixed physics step.
-     */
-    void StoreCurrentState();
-
-    /**
-     * @brief Read back the current position and rotation from Bullet after a step.
-     *
-     * Called internally by PhysicsSystem after each fixed physics step.
-     */
-    void UpdateCurrentState();
-
-  private:
-    /** @brief Descriptor capturing all creation-time parameters for this body. */
-    PhysicsBodyDesc m_desc;
-
-    /** @brief Underlying Bullet rigid body; owned and destroyed by the Bullet dynamics world. */
-    btRigidBody* m_bulletBody;
-
-    /**
-     * @brief Collision group bitmask for Bullet broadphase filtering.
-     *
-     * Default: 1 (bit 0 set — standard "world objects" group).
-     */
-    uint16_t m_collisionGroup = 1;
-
-    /**
-     * @brief Collision filter mask — defines which groups this body collides with.
-     *
-     * Default: 0xFFFF (all bits set — collide with every group).
-     */
-    uint16_t m_collisionMask = 0xFFFF;
-
-    // Interpolation state
-    XMFLOAT3 m_previousPosition{0, 0, 0};
-    XMFLOAT4 m_previousRotation{0, 0, 0, 1}; ///< Previous quaternion
-    XMFLOAT3 m_currentPosition{0, 0, 0};
-    XMFLOAT4 m_currentRotation{0, 0, 0, 1}; ///< Current quaternion
-};
-
-/**
- * @brief Physics constraint wrapper
- */
-class PhysicsConstraint
-{
-  public:
-    PhysicsConstraint(ConstraintType type, btTypedConstraint* bulletConstraint);
-    ~PhysicsConstraint();
-
-    ConstraintType GetType() const { return m_type; }
-    void SetEnabled(bool enabled);
-    bool IsEnabled() const;
-    void SetBreakingThreshold(float threshold);
-    float GetBreakingThreshold() const;
-
-    btTypedConstraint* GetBulletConstraint() const { return m_bulletConstraint; }
-
-  private:
-    ConstraintType m_type;
-    btTypedConstraint* m_bulletConstraint;
-};
 
 /**
  * @class PhysicsSystem
